// jobup-frontend/src/app/generated-sources/openapi/model/jobProposalDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface JobProposalDto { 
    id?: string;
    chatId?: string;
    senderId?: string;
    senderName?: string;
    senderType?: JobProposalDto.SenderTypeEnum;
    receiverId?: string;
    receiverName?: string;
    receiverType?: JobProposalDto.ReceiverTypeEnum;
    clientId?: string;
    workerId?: string;
    title?: string;
    description?: string;
    durationMinutes?: number;
    price?: number;
    location?: string;
    scheduledAt?: string;
    status?: JobProposalDto.StatusEnum;
    createdAt?: string;
    updatedAt?: string;
}
export namespace JobProposalDto {
    export type SenderTypeEnum = 'CLIENT' | 'WORKER';
    export const SenderTypeEnum = {
        Client: 'CLIENT' as SenderTypeEnum,
        Worker: 'WORKER' as SenderTypeEnum
    };
    export type ReceiverTypeEnum = 'CLIENT' | 'WORKER';
    export const ReceiverTypeEnum = {
        Client: 'CLIENT' as ReceiverTypeEnum,
        Worker: 'WORKER' as ReceiverTypeEnum
    };
    export type StatusEnum = 'PENDING' | 'ACCEPTED' | 'DECLINED' | 'NEGOTIATED';
    export const StatusEnum = {
        Pending: 'PENDING' as StatusEnum,
        Accepted: 'ACCEPTED' as StatusEnum,
        Declined: 'DECLINED' as StatusEnum,
        Negotiated: 'NEGOTIATED' as StatusEnum
    };
}